// WYNKWeb/src/app/Models/Refractionmaster.model.ts
export class Refraction {

  ID: number;
  IDOS: number;
  RegistrationTranID: number;
  UIN = '';
  Description: string;
  SubCategory: number = 0;
  Type: number;
  Ocular: string;
  OcularOS: string;
  DistSph: string = "";
  DistSphOS: string = "";
  NearCyl: string = "";
  NearCylOS: string = "";
  PinAxis: string = "";
  PinAxisOS: string = "";
  PowerGlass: string = "";
  PowerGlassOS: string = "";
  N_V_DESC: string = "";
  N_V_DESCOS: string = "";
  Add: string = "";
  Remarks: string = "";
  CreatedUTC: Date;
  UpdatedUTC: Date;
  CreatedBy: number;
  UpdatedBy: number;
  Iolnct: string;
  Iolat: string;
  Details: string = "";
  Amsler: string;
  Desc_Text: string;
  CV_normal: string;
  CV_defective: string;
  A_n_OD: boolean;
  A_abn_OD: boolean;
  A_n_OS: boolean;
  A_abn_OS: boolean;
  PD: string;
  MPDOD: string;
  MPDOS: string;
  Time: string;
  DV: number;
  NV: number;
  AddOS: string;
  DistSphNVOD: string;
  AddNVOD: string;
  DistSphNVOS: string;
  AddNVOS: string;
  ChartType: string;
}