// WYNKWeb/src/app/Models/PGPmodel.ts
export class PGP {

  ID: number;
  RegistrationTranID: number;
  Description: string;
  Ocular: string;
  OcularOS: string;
  DistSph: string;
  DistSphOS: string;
  NearCyl: string;
  NearCylOS: string;
  PinAxis: string;
  PinAxisOS: string;
  Add: string;
  AddOS: string;
  Details: string;
  CreatedUTC: Date;
  UpdatedUTC: Date;
  CreatedBy: number;
  UpdatedBy: number;
  OD: boolean;
  OS: boolean;
}