// WYNKWeb/src/app/Models/FINALPRESCRIPTION.model.ts
export class FINALPRESCRIPTION {

  ID: number;
  RegistrationTranID: number;
  Description: string;
  DV: number;
  NV: number;
  Ocular: string;
  OcularOS: string;
  DistSph: string;
  DistSphOS: string;
  NearCyl: string;
  NearCylOS: string;
  PinAxis: string;
  PinAxisOS: string;
  Add: string;
  AddOS: string;
  DistSphNVOD: string;
  AddNVOD: string;
  DistSphNVOS: string;
  AddNVOS: string;
  Remarks: string;
  CreatedUTC: Date;
  UpdatedUTC: Date;
  CreatedBy: number;
  UpdatedBy: number;
  PD: string;
  MPDOD: string;
  MPDOS: string;
  OD: false;
  OS: false;
}