// WYNKWeb/src/app/Models/Amsler.model.ts
export class Amsler {

  ID: number;
  RegistrationTranID: number;
  Description: string;
  Ocular: string;
  OcularOS: string;
  A_n_OD: boolean;
  A_abn_OD: boolean;
  A_n_OS: boolean;
  A_abn_OS: boolean;
  Desc_Text: string;
  Desc_TextOS: string;
  Date: Date;
  CreatedUTC: Date;
  UpdatedUTC: Date;
  CreatedBy: number;
  UpdatedBy: number;
  OD: boolean;
  OS: boolean;
}