// WYNKWeb/src/app/Models/ColorVision.model.ts
export class ColorVision {

  ID: number;
  RegistrationTranID: number;
  Description: string;
  Ocular: string;
  CV_normal: string;
  CV_defective: string;
  Desc_Text: string;
  OcularOS: string;
  CV_normalOS: string;
  CV_defectiveOS: string;
  Desc_TextOS: string;
  CreatedUTC: Date;
  UpdatedUTC: Date;
  CreatedBy: number;
  UpdatedBy: number;
  OD: boolean;
  OS: boolean;

}