// WYNKWeb/src/app/Models/VISUALACUITY.model.ts
export class VISUALACUITY {

  ID: number;
  RegistrationTranID: number;
  Description: string;
  SubCategory: number = 0;
  Ocular: string;
  OcularOS: string;
  DistSph: string;
  DistSphOS: string;
  NearCyl: string;
  NearCylOS: string;
  PinAxis: string;
  PinAxisOS: string;
  PowerGlass: string;
  PowerGlassOS: string;
  N_V_DESC: string;
  N_V_DESCOS: string;
  CreatedUTC: Date;
  UpdatedUTC: Date;
  CreatedBy: number;
  UpdatedBy: number;
  Remarks: string = "";
  ChartType: string;
}