// WYNKWeb/src/app/Models/REFRACTIONEXT.model.ts
export class REFRACTIONEXT {

  ID: number;
  RegistrationTranID: number;
  Description: string;
  SubCategory: number;
  Ocular: string;
  Sphdry: string;
  Cyldry: string;
  Axisdry: string;
  Sphwet: string;
  Cylwet: string;
  Axiswet: string;
  OcularOS: string;
  Sphdryos: string;
  Cyldryos: string;
  Axisdryos: string;
  Sphwetos: string;
  Cylwetos: string;
  Axiswetos: string;
  Date: Date;
  CreatedUTC: Date;
  UpdatedUTC: Date;
  CreatedBy: number;
  UpdatedBy: number;
}